client/quizlib: slice when merging, don't splice
currentQ.splice() will chop the unsynced answer off the end of
currentQ. This isn't a useful side-effect, since we should be wholesale
replacing the queue later on.

In combination with 'force-sync' generating 2 syncs (previous commit),
this was resulting in-progress questions getting lost when syncs were
happening, so if a student was particularly keen on pressing "New
question", their question would be overwritten with the older one,
meaning questions would appear to not have a correct answer (the new
question is marked as-per the previous).

Ideally this code would be ditched entirely in preference to doing this
server-side, but that's an interesting upgrade to do.